# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * PackSMP Bridge API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminCredentials {
  password: string;
}

export interface PasswordChange {
  currentPassword: string;
  newPassword: string;
}

export interface AuthToken {
  token: string;
  expiresIn: string;
}

export interface AdminUser {
  role: string;
  authenticated: boolean;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ConfigEntry {
  key: string;
  value: string;
}

export interface MaintenanceState {
  enabled: boolean;
}

export interface RconRequest {
  command: string;
}

export interface RconResult {
  success: boolean;
  /** @nullable */
  output?: string | null;
}

export type DashboardStatsDiscord = {
  connected: boolean;
  /** @nullable */
  username?: string | null;
  guilds: number;
};

export type DashboardStatsMinecraft = {
  connected: boolean;
  username: string;
  host: string;
  port: number;
};

export type DashboardStatsRcon = {
  connected: boolean;
  host: string;
  port: number;
};

export type DashboardStatsSync = {
  enabled: boolean;
  maintenance: boolean;
};

export type DashboardStatsStats = {
  verifiedAccounts: number;
  totalErrors: number;
  pendingMessages: number;
  uptimeSeconds: number;
};

export interface LogEntry {
  id: string;
  level: string;
  service: string;
  message: string;
  /** @nullable */
  context?: string | null;
  createdAt: string;
}

export interface DashboardStats {
  discord: DashboardStatsDiscord;
  minecraft: DashboardStatsMinecraft;
  rcon: DashboardStatsRcon;
  sync: DashboardStatsSync;
  stats: DashboardStatsStats;
  recentErrors: LogEntry[];
}

export interface VerificationEntry {
  id: string;
  discordId: string;
  /** @nullable */
  mcNick?: string | null;
  /** @nullable */
  mcUuid?: string | null;
  isVerified: boolean;
  /** @nullable */
  verifiedAt?: string | null;
  createdAt: string;
}

export interface VerificationList {
  rows: VerificationEntry[];
  total: number;
  page: number;
  limit: number;
}

export interface MessageEntry {
  id: string;
  fromType: string;
  fromDisplay: string;
  toId: string;
  message: string;
  isDelivered: boolean;
  /** @nullable */
  deliveredAt?: string | null;
  createdAt: string;
}

export interface MessageList {
  rows: MessageEntry[];
  total: number;
}

export type SystemConfigDbItem = {
  key: string;
  value: string;
  updatedAt: string;
};

export type SystemConfigEnv = {[key: string]: string};

export type SystemConfigRuntime = {
  syncEnabled: boolean;
  maintenanceMode: boolean;
};

export interface SystemConfig {
  db: SystemConfigDbItem[];
  env: SystemConfigEnv;
  runtime: SystemConfigRuntime;
}

export interface LogList {
  rows: LogEntry[];
  total: number;
}

export interface ChatLogEntry {
  id: string;
  source: string;
  author: string;
  /** @nullable */
  authorId?: string | null;
  message: string;
  createdAt: string;
}

export interface ChatLogList {
  rows: ChatLogEntry[];
  total: number;
}

export type SystemStatusDiscord = {
  connected: boolean;
  /** @nullable */
  username?: string | null;
  guilds: number;
};

export type SystemStatusMinecraft = {
  connected: boolean;
  username: string;
  host: string;
  port: number;
};

export type SystemStatusRcon = {
  connected: boolean;
};

export type SystemStatusSync = {
  enabled: boolean;
  maintenance: boolean;
};

export interface SystemStatus {
  discord: SystemStatusDiscord;
  minecraft: SystemStatusMinecraft;
  rcon: SystemStatusRcon;
  sync: SystemStatusSync;
  uptime: number;
  timestamp: string;
}

export type GetVerificationsParams = {
page?: number;
limit?: number;
search?: string;
};

export type GetMessagesParams = {
limit?: number;
};

export type GetSystemLogsParams = {
limit?: number;
level?: string;
};

export type GetChatLogsParams = {
limit?: number;
source?: string;
};